// src/domain/enums/PrinterStatus.ts
export enum PrinterStatus {
  SUCCESS = 'SUCCESS',
  PENDING = 'PENDING',
  OUT_OF_PAPER = 'OUT_OF_PAPER',
  COVER_OPEN = 'COVER_OPEN',
  LOW_BATTERY = 'LOW_BATTERY',
  HARDWARE_ERROR = 'HARDWARE_ERROR',
  INVALID_PARAM = 'INVALID_PARAM',
  UNKNOWN_ERROR = 'UNKNOWN_ERROR',
  MODULE_NOT_AVAILABLE = 'MODULE_NOT_AVAILABLE',
}